requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gps3==0.33.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
mailersend==0.5.8
matplotlib-inline==0.1.7
maxminddb==2.6.2
multidict==6.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyJWT==1.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
ratelim==0.1.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.3.0
sib-api-v3-sdk==7.6.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
user-agents==2.2.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.17.1
